add evidence profile component
